test: move record to the propert entity scenarion
